requirements.txt
# Core Flask dependencies
Flask==3.1.1
flask-cors==6.0.1
Werkzeug==3.1.3

# Database & Auth
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
bcrypt==4.2.0

# PDF processing
pdfplumber==0.11.7
PyMuPDF==1.26.3
pdfminer.six==20250506

# Document processing
python-docx==1.1.2
pytesseract==0.3.13

# AI — dual-provider (set AI_PROVIDER=gemini or openai in .env)
google-generativeai>=0.8.0
openai==1.95.1

# Environment and configuration
python-dotenv==1.1.1

# Utilities
requests==2.32.4
Pillow==11.3.0

# Vector store
chromadb>=1.0.0,<2.0.0

# LangChain RAG pipeline
langchain==0.3.25
langchain-core>=0.3.58
langchain-text-splitters==0.3.8
langchain-openai==0.3.12
langchain-google-genai==2.1.4
langchain-chroma==0.2.4

# MCP (Model Context Protocol)
mcp>=1.0.0

# Async task queue
celery[redis]==5.4.0
redis==5.2.1

# Rate limiting (distributed via Redis)
Flask-Limiter==3.12

# Structured logging
structlog==25.1.0

# AWS S3 (optional file storage)
boto3==1.38.0

# Production server
gunicorn==23.0.0
